+++ 4HC3 Assignment 2 - Code/HC3 A2/Transfer3.xaml.cs
@@ -56,28 +56,28 @@ namespace HC3_A2
             string account2 = toAccount.Split(' ')[0].ToLower();
             if (account1 == "chequing")
             {
-                balance1 = Convert.ToString(Convert.ToInt32(balance1) - Convert.ToInt32(amount.Substring(2)));
+                balance1 = Convert.ToString(Convert.ToDouble(balance1) - Convert.ToDouble(amount.Substring(2)));
             }
             else if (account1 == "savings")
             {
-                balance2 = Convert.ToString(Convert.ToInt32(balance2) - Convert.ToInt32(amount.Substring(2)));
+                balance2 = Convert.ToString(Convert.ToDouble(balance2) - Convert.ToDouble(amount.Substring(2)));
             }
             else if (account1 == "other")
             {
-                balance3 = Convert.ToString(Convert.ToInt32(balance3) - Convert.ToInt32(amount.Substring(2)));
+                balance3 = Convert.ToString(Convert.ToDouble(balance3) - Convert.ToDouble(amount.Substring(2)));
             }
 
             if (account2 == "chequing")
             {
-                balance1 = Convert.ToString(Convert.ToInt32(balance1) + Convert.ToInt32(amount.Substring(2)));
+                balance1 = Convert.ToString(Convert.ToDouble(balance1) + Convert.ToDouble(amount.Substring(2)));
             }
             else if (account2 == "savings")
             {
-                balance2 = Convert.ToString(Convert.ToInt32(balance2) + Convert.ToInt32(amount.Substring(2)));
+                balance2 = Convert.ToString(Convert.ToDouble(balance2) + Convert.ToDouble(amount.Substring(2)));
             }
             else if (account2 == "other")
             {
-                balance3 = Convert.ToString(Convert.ToInt32(balance3) + Convert.ToInt32(amount.Substring(2)));
+                balance3 = Convert.ToString(Convert.ToDouble(balance3) + Convert.ToDouble(amount.Substring(2)));
             }
 
             file.WriteLine(balance1);
